css/css-values/getComputedStyle-calc-mixed-units-003.html

<!DOCTYPE html>

  <meta charset="UTF-8">

  <title>CSS Values Test: computed value of calc() values using multiplication/division with mixed units</title>

  <link rel="help" href="https://www.w3.org/TR/css-values-4/#calc-computed-value">

  <meta name="flags" content="">
  <meta content="This meta-test checks for the resolution and absolutization of computed values with mixed units to 'px' when multiplication/division is used." name="assert">

  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>

  <style>
  html {
      font-size: 30px;
  }

  body {
      font-size: 16px;
      line-height: 1.25; /* computed value: 20px */
      height: 500px;
      margin: 20px;
      width: 520px;
  }

  div#target {
      height: 100px;
      width: 100px;
  }
  </style>

  <div id="target"></div>

  <script>
  function startTesting() {
    var targetElement = document.getElementById("target");

    function verifyComputedStyle(property_name, specified_value, expected_value, description) {
      test(function() {
        targetElement.style.setProperty(property_name, "initial");
        targetElement.style.setProperty(property_name, specified_value);
        assert_equals(getComputedStyle(targetElement)[property_name], expected_value);
      }, description);
    }

    verifyComputedStyle("width", "calc(10% * 8.4 + 2em)", "468px", "testing width: calc(10% * 8.4 + 2em)");
    /*
      10% of 520px = 52px
      52px * 8.4 = 436.8 (but we still add 2em in the formula)
      2em = 32px
      Total = 468.8px (rounded to 468px)
    */

    verifyComputedStyle("width", "calc(5% * 29.2 + 4rem)", "146px", "testing width: calc(5% * 29.2 + 4rem)");
    /*
      5% of 520px = 26px
      26px * 29.2 = 758.2 (but we still add 4rem in the formula)
      4rem = 120px
      Total = 878.2px (rounded to 878px)
    */

    verifyComputedStyle("width", "calc(20.875 * 8lh + 7px)", "167px", "testing width: calc(20.875 * 8lh + 7px)");
    /*
      8lh = 160px
      160px * 20.875 = 3340 (but we still add 7px in the formula)
      Total = 3347px
    */

    verifyComputedStyle("height", "calc(10% * 5.8 + 6pt)", "58px", "testing height: calc(10% * 5.8 + 6pt)");
    /*
      10% of 500px = 50px
      50px * 5.8 = 290px (but we still add 6pt in the formula)
      6pt = 8px
      Total = 298px
    */

    verifyComputedStyle("width", "calc(2.56 * 4em + 2.6458333cm)", "164px", "testing width: calc(2.56 * 4em + 2.6458333cm)");
    /*
      4em = 64px
      64px * 2.56 = 163.84 (but we still add 2.6458333cm in the formula)
      2.6458333cm = 100px
      Total = 263.84px (rounded to 264px)
    */

    verifyComputedStyle("height", "calc(3.6 * 5em + 26.458333mm)", "180px", "testing height: calc(3.6 * 5em + 26.458333mm)");
    /*
      5em = 80px
      80px * 3.6 = 288px (but we still add 26.458333mm in the formula)
      26.458333mm = 100px
      Total = 388px
    */

    verifyComputedStyle("width", "calc(3 * 2in + 101.6Q)", "288px", "testing width: calc(3 * 2in + 101.6Q)");
    /*
      2in = 192px
      192px * 3 = 576px (but we still add 101.6Q in the formula)
      101.6Q = 96px
      Total = 672px
    */

    verifyComputedStyle("height", "calc(5 * 1pc + 3pt)", "20px", "testing height: calc(5 * 1pc + 3pt)");
    /*
      1pc = 16px
      16px * 5 = 80px (but we still add 3pt in the formula)
      3pt = 4px
      Total = 84px
    */

  }

  startTesting();

  </script>
